Move ButtonGroup to its own folder
