+++ src/components/button/index.js
@@ -1,5 +1,4 @@
 import Button from './Button';
-import ButtonGroup from './ButtonGroup';
 
 export default Button;
-export { Button, ButtonGroup };
+export { Button };
