+++ src/components/buttonGroup/buttonGroup.stories.js
@@ -3,7 +3,7 @@ import { addStoryInGroup, LOW_LEVEL_BLOCKS } from '../../../.storybook/utils';
 import { boolean } from '@storybook/addon-knobs/react';
 import { Store, State } from '@sambego/storybook-state';
 import { IconAddMediumOutline } from '@teamleader/ui-icons';
-import Button from './Button';
+import Button from '../button';
 import ButtonGroup from './ButtonGroup';
 
 const store = new Store({
