+++ src/components/buttonGroup/index.js
@@ -1,0 +1,4 @@
+import ButtonGroup from './ButtonGroup';
+
+export default ButtonGroup;
+export { ButtonGroup };
