+++ src/components/overviewPage/overviewPage.stories.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { text } from '@storybook/addon-knobs/react';
 import { addStoryInGroup, COMPOSITIONS } from '../../../.storybook/utils';
 import { IconChevronDownSmallOutline } from '@teamleader/ui-icons';
-import Button, { ButtonGroup } from '../button';
+import Button from '../button';
+import ButtonGroup from '../buttonGroup';
 import OverviewPage from './OverviewPage';
 import { COLOR } from '../../constants';
 import { TextBody } from '../typography';
